server/requirements.txt
alembic==1.14.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
docker==7.1.0
double-elimination==1.4.1
filelock==3.17.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fsspec==2024.12.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.37
sympy==1.13.1
torch==2.5.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
